+++ galsim/config/noise.py
@@ -184,7 +184,7 @@ def GetSky(config, base):
             return sky_level * wcs.pixelArea(base['image_pos'])
         else:
             # This calculation is non-trivial, so store this in case we need it again.
-            tag = (base['file_num'], base['image_num'])
+            tag = (id(base), base['file_num'], base['image_num'])
             if config.get('current_sky_tag',None) == tag:
                 return config['current_sky']
             else:
@@ -526,18 +526,17 @@ class COSMOSNoiseBuilder(NoiseBuilder):
 
     def getCOSMOSNoise(self, config, base, rng=None):
         # Save the constructed CorrelatedNoise object, since we might need it again.
-        tag = (base['file_num'], base['image_num'])
+        tag = (id(base), base['file_num'], base['image_num'])
         if self.current_cn_tag == tag:
             return self.current_cn
         else:
-            req = { 'file_name' : str }
-            opt = { 'cosmos_scale' : float, 'variance' : float }
+            opt = { 'file_name' : str, 'cosmos_scale' : float, 'variance' : float }
 
-            kwargs = galsim.config.GetAllParams(config, base, req=req, opt=opt,
+            kwargs = galsim.config.GetAllParams(config, base, opt=opt,
                                                 ignore=noise_ignore)[0]
             if rng is None:
                 rng = base['rng']
-            cn = galsim.correlatednoise.getCOSMOSNoise(rng, **kwargs)
+            cn = galsim.correlatednoise.getCOSMOSNoise(rng=rng, **kwargs)
             self.current_cn = cn
             self.current_cn_tag = tag
             return cn